Add balance option to make-random-forest
The balance option allows for equal sampling among the outcome of the
objective variable. This is really only useful when one outcome is
over-represented in the data. When enabled this option should improve
the true-positive and false-positive rates at the expense of losing some
information from the undersampled outcome and increasing bias in the
oversampled outcome.